Moving code for data generator to CartPole folder
